Fix AOM JSON with official BMM (#372)
Make JSON serialization of AOM, plus the AOM itself more in line with the spec.

In the JSON (+yaml and ODIN) mapping of the AOM:

adds is_ prefixes to boolean fields
renames patterned_constraint to pattern_constraint
still accepts the old json
configurable to switch back to old behaviour
API changes, separately from the JSON:

Renamed RMJacksonConfiguration to ArchieJacksonConfiguration because of changed scope
added archetypeBooleanIsPrefix and addPatternConstraintTypo to ArchieJacksonConfiguration, set to false and true respectively to fall back to the old behaviour
rename patternedConstraint in CTemporal to patternConstraint
added lots of @nullable annotations in the AOM
And:

adds a test that runs the automated AOM/BMM comparison
This test still fails, needs work to check the differences and add a list of known differences.

Breaking changes:

ArchieRMJacksonConfiguration is renamed to ArchieJacksonConfiguration
JacksonUtil.getObjectMapper() and configureObjectMapper now returns a standards compliant version of the ObjectMapper
the constructor of ArchieJacksonConfiguration is now marked private. Before it created a legacy version of the configuration. That can still be done by calling ArchieJacksonConfiguration.createLegacyConfiguration()
the static methodArchieJacksonConfiguration.createLegacyConfiguration() has been added in order to generate output compatible with earlier versions of Archie.
odin to json converter now outputs _type instead of @type - should not matter for most people.
the default for the ExampleJsonInstanceGenerator is now to generate _type instead of @type
the json output for Operator, even the legacy output, no longer contains left_operand and right_operand, unless it is an instance of BinaryOperator. This can break other apps, but output from old archie versions should parse just fine with the newer ones, and legacy output from this archie should still parse with older ones because it is serialized twice in old versions.
Migration notes:

even for standards compliant json, always including the type name in the json for every object could be useful!