Appy Commit 038d746d from openehr/archie (simplify_rm branch).
Make MetaModels available to lower levels of ADLParser
Use MetaModels during parsing of CPrimitiveObjects to get access to the correct ModelNamingStrategy. Currently conditional since ADLParser can be created with no MetaModels as well.

In PrimitivesConstraintParser, if the ModelNamingStrategy object can be found, write in the correct RmTypeNames into CPrimitiveObjects (CTerminologyCode and so on).

Rename ArchieModelNamingStrategy to OpenEhrModelNamingStrategy.

branches simplify_rm, origin/simplify_rm